+++ test/plugin/infer_extension_test.rb
@@ -1,0 +1,88 @@
+require "test_helper"
+require "shrine/plugins/infer_extension"
+
+describe Shrine::Plugins::InferExtension do
+  before do
+    @uploader = uploader { plugin :infer_extension }
+    @shrine = @uploader.class
+  end
+
+  describe ":rack_mime analyzer" do
+    before do
+      @shrine.plugin :infer_extension, inferrer: :rack_mime
+    end
+
+    it "determines extension from MIME type" do
+      assert_equal ".png", @shrine.infer_extension("image/png")
+    end
+
+    it "chooses one extension from available extensions" do
+      assert_equal ".jpeg", @shrine.infer_extension("image/jpeg")
+    end
+
+    it "returns empty string when it couldn't determine extension" do
+      assert_nil @shrine.infer_extension("foo")
+    end
+  end
+
+  describe ":mime_types analyzer" do
+    before do
+      @shrine.plugin :infer_extension, inferrer: :mime_types
+    end
+
+    it "determines extension from MIME type" do
+      assert_equal ".png", @shrine.infer_extension("image/png")
+    end
+
+    it "chooses one extension from available extensions" do
+      assert_equal ".jpeg", @shrine.infer_extension("image/jpeg")
+    end
+
+    it "returns empty string when it couldn't determine extension" do
+      assert_nil @shrine.infer_extension("foo")
+    end
+  end
+
+  describe ":mini_mime analyzer" do
+    before do
+      @shrine.plugin :infer_extension, inferrer: :mini_mime
+    end
+
+    it "determines extension from MIME type" do
+      assert_equal ".png", @shrine.infer_extension("image/png")
+    end
+
+    it "chooses one extension from available extensions" do
+      assert_equal ".jpe", @shrine.infer_extension("image/jpeg")
+    end
+
+    it "returns empty string when it couldn't determine extension" do
+      assert_nil @shrine.infer_extension("foo")
+    end
+  end
+
+  it "has a default inferrer" do
+    assert_equal ".png", @shrine.infer_extension("image/png")
+  end
+
+  it "allows passing a custom inferrer" do
+    @shrine.plugin :infer_extension, inferrer: ->(mime_type) { ".foo" }
+    assert_equal ".foo", @shrine.infer_extension("image/jpeg")
+
+    @shrine.plugin :infer_extension, inferrer: ->(mime_type, inferrers) { inferrers[:rack_mime].call(mime_type) }
+    assert_equal ".jpeg", @shrine.infer_extension("image/jpeg")
+  end
+
+  it "automatically infers extension when generating location" do
+    uploaded_file = @uploader.upload(fakeio(content_type: "image/jpeg"))
+    assert_equal ".jpeg", File.extname(uploaded_file.id)
+    assert_nil uploaded_file.original_filename
+  end
+
+  it "provides access to extension inferrers" do
+    inferrers = @shrine.extension_inferrers
+
+    assert_equal ".jpeg", inferrers[:rack_mime].call("image/jpeg")
+    assert_equal ".jpeg", inferrers[:mime_types].call("image/jpeg")
+  end
+end